remove models folder
