+++ components/[pageId]/DocumentPage/components/BountyProperties/BountyProperties.tsx
@@ -19,11 +19,10 @@ import { usePaymentMethods } from 'hooks/usePaymentMethods';
 import { useUser } from 'hooks/useUser';
 import type { ApplicationWithTransactions } from 'lib/applications/interfaces';
 import type { AssignedBountyPermissions, BountyPermissions, UpdateableBountyFields } from 'lib/bounties';
-import type { BountyCreationData } from 'lib/bounties/interfaces';
+import type { BountyCreationData, BountyWithDetails } from 'lib/bounties/interfaces';
 import type { TargetPermissionGroup } from 'lib/permissions/interfaces';
 import debouncePromise from 'lib/utilities/debouncePromise';
 import { isTruthy } from 'lib/utilities/types';
-import type { BountyWithDetails } from 'models';
 import { useRouter } from 'next/router';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import BountyApplicantForm from './components/BountyApplicantForm';
