+++ components/[pageId]/DocumentPage/components/BountyProperties/components/BountyApplicantsTable/BountyApplicantTableRow.tsx
@@ -17,9 +17,8 @@ import { useContributors } from 'hooks/useContributors';
 import { useUser } from 'hooks/useUser';
 import type { ApplicationWithTransactions } from 'lib/applications/actions';
 import type { ReviewDecision, SubmissionReview } from 'lib/applications/interfaces';
-import type { AssignedBountyPermissions } from 'lib/bounties/interfaces';
+import type { AssignedBountyPermissions, BountyWithDetails } from 'lib/bounties/interfaces';
 import { humanFriendlyDate } from 'lib/utilities/dates';
-import type { BountyWithDetails } from 'models';
 import { useEffect, useState } from 'react';
 import charmClient from 'charmClient';
 import { shortenHex } from 'lib/utilities/strings';
