+++ components/[pageId]/DocumentPage/components/BountyProperties/components/BountyApplicantsTable/BountyApplicantsTable.tsx
@@ -15,9 +15,8 @@ import { useBounties } from 'hooks/useBounties';
 import { useUser } from 'hooks/useUser';
 import type { ApplicationWithTransactions } from 'lib/applications/actions';
 import { countValidSubmissions, submissionsCapReached as submissionsCapReachedFn } from 'lib/applications/shared';
-import type { AssignedBountyPermissions } from 'lib/bounties/interfaces';
+import type { AssignedBountyPermissions, BountyWithDetails } from 'lib/bounties/interfaces';
 import { isBountyLockable } from 'lib/bounties/shared';
-import type { BountyWithDetails } from 'models';
 import { useEffect, useState } from 'react';
 import BountyApplicantTableRow from './BountyApplicantTableRow';
