+++ components/bounties/components/BountyCard.tsx
@@ -1,7 +1,7 @@
 import { Box, CardHeader, Typography } from '@mui/material';
 import type { Page } from '@prisma/client';
 import { fancyTrim } from 'lib/utilities/strings';
-import type { BountyWithDetails } from 'models';
+import type { BountyWithDetails } from 'lib/bounties/interfaces';
 import { memo } from 'react';
 import BountyStatusBadge from './BountyStatusBadge';
