+++ lib/bounties/listAvailableBounties.ts
@@ -1,5 +1,5 @@
 import { prisma } from 'db';
-import type { BountyWithDetails } from 'models';
+import type { BountyWithDetails } from 'lib/bounties/interfaces';
 import { hasAccessToSpace } from 'lib/middleware';
 import { accessiblePagesByPermissionsQuery, includePagePermissions } from 'lib/pages/server';
 import type { AvailableResourcesRequest } from '../permissions/interfaces';
