+++ lib/bounties/reviewBountySuggestion.ts
@@ -1,7 +1,7 @@
 import { prisma } from 'db';
 import { includePagePermissions } from 'lib/pages/server';
 import { WrongStateError } from 'lib/utilities/errors';
-import type { BountyWithDetails } from 'models';
+import type { BountyWithDetails } from 'lib/bounties/interfaces';
 import { getBountyOrThrow } from './getBounty';
 import type { SuggestionAction, SuggestionApproveAction, SuggestionRejectAction } from './interfaces';
