+++ lib/bounties/updateBountySettings.ts
@@ -1,6 +1,6 @@
 import { prisma } from 'db';
 import { DataNotFoundError, InvalidInputError, PositiveNumbersOnlyError } from 'lib/utilities/errors';
-import type { BountyWithDetails } from 'models';
+import type { BountyWithDetails } from 'lib/bounties/interfaces';
 import { countValidSubmissions } from '../applications/shared';
 import { setBountyPermissions } from '../permissions/bounties';
 import { getBountyOrThrow } from './getBounty';
