+++ lib/permissions/__tests__/compareBountyPagePermissions.spec.ts
@@ -1,6 +1,6 @@
 import { getBounty } from 'lib/bounties';
 import { assignRole, listRoleMembers } from 'lib/roles';
-import type { BountyWithDetails } from 'models';
+import type { BountyWithDetails } from 'lib/bounties/interfaces';
 import { generateBounty, generateRole, generateSpaceUser, generateUserAndSpaceWithApiToken } from 'testing/setupDatabase';
 import { queryBountyPermissions } from '../bounties';
 import { compareBountyPagePermissions } from '../compareBountyPagePermissions';
