+++ models/index.ts
@@ -2,4 +2,3 @@
 export * from './User';
 export * from './Page';
 export * from './Bounty';
-export * from './Task';
