+++ testing/setupDatabase.ts
@@ -10,7 +10,8 @@ import type { ProposalReviewerInput } from 'lib/proposal/interface';
 import { syncProposalPermissions } from 'lib/proposal/syncProposalPermissions';
 import { createUserFromWallet } from 'lib/users/createUser';
 import { typedKeys } from 'lib/utilities/objects';
-import type { BountyWithDetails, LoggedInUser } from 'models';
+import type { LoggedInUser } from 'models';
+import type { BountyWithDetails } from 'lib/bounties/interfaces';
 import { IDENTITY_TYPES } from 'models';
 import { v4 } from 'uuid';
 import { boardWithCardsArgs } from './generate-board-stub';
